move PreparedRequest into a separate module
