+++ libindy_vdr/src/ledger/mod.rs
@@ -2,6 +2,8 @@
 pub mod constants;
 /// Identifiers for stored objects on the ledger
 pub mod identifiers;
+/// Assembled ledger transaction request
+mod prepared_request;
 /// Types for constructing ledger transactions
 #[macro_use]
 pub mod requests;
@@ -9,5 +11,6 @@ pub mod requests;
 /// Helpers for constructing ledger requests
 mod request_builder;
 
-pub use request_builder::{PreparedRequest, RequestBuilder};
+pub use prepared_request::PreparedRequest;
+pub use request_builder::RequestBuilder;
 pub(crate) use requests::author_agreement::TxnAuthrAgrmtAcceptanceData;
